% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normic_colors.R
\docType{data}
\name{normic_colors}
\alias{normic_colors}
\title{Nordic microbes A/S official colors}
\format{
An object of class \code{list} of length 9.
}
\usage{
normic_colors
}
\description{
Nordic microbes A/S official colors
}
\keyword{datasets}
