% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normic_palettes.R
\name{normic_palettes}
\alias{normic_palettes}
\title{Create palette from normic_colors}
\usage{
normic_palettes(
  palette,
  n,
  all_palettes = normic_colors,
  type = c("discrete", "continuous")
)
}
\arguments{
\item{palette}{A palette from the normic_colors vector}

\item{n}{Number of colors}

\item{all_palettes}{normic_colors}

\item{type}{one of "discrete" or "continuous"}
}
\value{
A palette object
}
\description{
Create palette from normic_colors
}
\examples{
normic_palettes("greens")
}
